+++ test/vectorint1/popintvector.cc
@@ -12,7 +12,7 @@ void POPintVector::SetInternalData(vector<int> d)
 
 void POPintVector::ShowPOPintVector()
 {
-	printf("VectorInt1: Vector is of size %d: ", theData.size());
+	printf("VectorInt1: Vector is of size %d: ", static_cast<int>(theData.size()));
 	for (long i=0; i<theData.size(); i++) printf("%d ",theData[i]);
 	printf("\nVectorInt1: MyData = %d\n", myData);
 }
@@ -28,23 +28,18 @@ void POPintVector::Serialize(POPBuffer &buf, bool pack)
 	int a;
 	long size;
 	vector<int>::iterator iter;
-	if (pack)
-	{
-		size=theData.size();
+	if (pack) {
+		size = static_cast<int>(theData.size());
 		buf.Pack(&size,1);
-		for (iter=theData.begin(); iter!=theData.end(); iter++)
-		{
+		for (iter = theData.begin(); iter != theData.end(); iter++) {
 			a = *iter;
-			buf.Pack(&a,1);
+			buf.Pack(&a, 1);
 		}
-		buf.Pack(&myData,1);
-	}
-	else
-	{
-		buf.UnPack(&size,1);
+		buf.Pack(&myData, 1);
+	} else {
+		buf.UnPack(&size, 1);
 		theData.clear();
-		for (long i=0; i<size; i++)
-		{
+		for (long i = 0; i < static_cast<int>(size); i++) {
 			buf.UnPack(&a,1);
 			theData.push_back(a);
 		}
